Support connection-less issue-credential and present-proof
To solve the issue #108 If Agent receives out-of-band message with attachment without handshake protocols, It precess the attachment and send message to counter party if necessary. When It send a message, It use the service decorator in the out-of-band message. By doing this Agent supports connection-less present-proof.

Sharing the main concept. unit test is not implemented yet.