Support connection-less credential exchange and proof exchange.
To solve the issue #108
If Agent receives out-of-band message with attachment without handshake protocols, It precess the attachment and send message to counter party if necessary. When It send a message, It use the service decorator in the out-of-band message. By doing this Agent supports connection-less credential exchange and proof exchange.

Second approach. unit test is not implemented yet.

Signed-off-by: kukgini <kukgini@gmail.com>